Moving functions out of huge tok_geo.c file. Low level functions are now in private header. Going to try to move functions related to finding turning points out of main tok_geo.c file so code is more readable
